+++ src/main/java/com/phantomwing/rusticdelight/datagen/ModItemTagsProvider.java
@@ -8,6 +8,7 @@ import com.phantomwing.rusticdelight.tags.ModTags;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.tags.ItemTagsProvider;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.world.item.Items;
@@ -38,7 +39,21 @@ public class ModItemTagsProvider extends ItemTagsProvider {
     private void addModTags() {
         this.tag(ModTags.Items.COOKING_OIL_INGREDIENTS).add(
                 ModItems.COTTON_SEEDS.get()
+        ).addOptional(ResourceLocation.fromNamespaceAndPath(CompatibilityTags.FRYCOOKS_DELIGHT, "canola_seeds"));
+
+        this.tag(ModTags.Items.COOKING_OIL).add(
+                ModItems.COOKING_OIL.get()
+        ).addOptional(ResourceLocation.fromNamespaceAndPath(CompatibilityTags.FRYCOOKS_DELIGHT, "canola_oil"));
+
+        this.tag(ModTags.Items.CHERRY_BLOSSOM_INGREDIENTS).add(
+                Items.PINK_PETALS,
+                Items.CHERRY_LEAVES,
+                Items.CHERRY_SAPLING
         );
+
+        this.tag(ModTags.Items.CALAMARI_ROLL_INGREDIENTS).add(
+                ModItems.CALAMARI_SLICE.get()
+        ).addOptional(ResourceLocation.fromNamespaceAndPath(CompatibilityTags.MINERS_DELIGHT, "tentacles"));
     }
 
     private void addMinecraftTags() {
@@ -107,10 +122,18 @@ public class ModItemTagsProvider extends ItemTagsProvider {
         this.tag(CommonTags.FOODS_BELL_PEPPER).addTag(
                 CommonTags.CROPS_BELL_PEPPER
         );
-        this.tag(CommonTags.FOODS_RAW_CALAMARI).add(
-                ModItems.CALAMARI.get(),
-                ModItems.CALAMARI_SLICE.get()
-        );
+        this.tag(CommonTags.FOODS_RAW_CALAMARI)
+                .add(
+                    ModItems.CALAMARI.get(),
+                    ModItems.CALAMARI_SLICE.get()
+                )
+                .addOptional(ResourceLocation.fromNamespaceAndPath(CompatibilityTags.CULTURAL_DELIGHTS,"squid"))
+                .addOptional(ResourceLocation.fromNamespaceAndPath(CompatibilityTags.CULTURAL_DELIGHTS,"glow_squid"))
+                .addOptional(ResourceLocation.fromNamespaceAndPath(CompatibilityTags.CULTURAL_DELIGHTS,"raw_calamari"))
+                .addOptional(ResourceLocation.fromNamespaceAndPath(CompatibilityTags.MINERS_DELIGHT,"squid"))
+                .addOptional(ResourceLocation.fromNamespaceAndPath(CompatibilityTags.MINERS_DELIGHT,"glow_squid"))
+                .addOptional(ResourceLocation.fromNamespaceAndPath(CompatibilityTags.MINERS_DELIGHT,"tentacles"));
+
         this.tag(CommonTags.FOODS_COOKED_CALAMARI).add(
                 ModItems.COOKED_CALAMARI.get(),
                 ModItems.COOKED_CALAMARI_SLICE.get()
@@ -122,12 +145,30 @@ public class ModItemTagsProvider extends ItemTagsProvider {
     }
 
     private void addCompatibilityTags() {
+        // Create
         this.tag(CompatibilityTags.CREATE_UPRIGHT_ON_BELT).add(
                 ModItems.COOKING_OIL.get()
         );
 
+        // Farmer's Delight
         this.tag(vectorwing.farmersdelight.common.tag.ModTags.CABBAGE_ROLL_INGREDIENTS).add(
                 ModItems.POTATO_SLICES.get()
+        );
+
+        // Miner's Delight
+        this.tag(CompatibilityTags.MINERS_DELIGHT_TENTACLES).add(
+                ModItems.CALAMARI.get(),
+                ModItems.CALAMARI_SLICE.get(),
+                ModItems.COOKED_CALAMARI.get(),
+                ModItems.COOKED_CALAMARI_SLICE.get()
+        );
+
+        // Frycook's Delight
+        this.tag(CompatibilityTags.FISH_SLICES).add(
+                ModItems.CALAMARI_SLICE.get()
+        );
+        this.tag(CompatibilityTags.HAS_FISH_SLICE).add(
+                ModItems.CALAMARI.get()
         );
     }
 }