+++ src/main/java/com/phantomwing/rusticdelight/tags/CompatibilityTags.java
@@ -6,6 +6,7 @@ import net.minecraft.tags.ItemTags;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
+import vectorwing.farmersdelight.FarmersDelight;
 
 /**
  * References to tags belonging to other mods, which Farmer's Delight innately supports.
@@ -13,6 +14,9 @@ import net.minecraft.world.level.block.Block;
  */
 public class CompatibilityTags
 {
+    public static final String FORGE = "forge";
+    public static final String FARMERS_DELIGHT = FarmersDelight.MODID;
+
     // Create
     public static final String CREATE = "create";
     public static final TagKey<Item> CREATE_UPRIGHT_ON_BELT = externalItemTag(CREATE, "upright_on_belt");
@@ -28,6 +32,18 @@ public class CompatibilityTags
     public static final TagKey<Item> SERENE_SEASONS_SPRING_CROPS = externalItemTag(SERENE_SEASONS, "spring_crops");
     public static final TagKey<Item> SERENE_SEASONS_SUMMER_CROPS = externalItemTag(SERENE_SEASONS, "summer_crops");
     public static final TagKey<Item> SERENE_SEASONS_WINTER_CROPS = externalItemTag(SERENE_SEASONS, "winter_crops");
+
+    // Cultural Delights
+    public static final String CULTURAL_DELIGHTS = "culturaldelights";
+
+    // Miner's Delight
+    public static final String MINERS_DELIGHT = "miners_delight";
+    public static final TagKey<Item> MINERS_DELIGHT_TENTACLES = externalItemTag(FORGE, "tentacles");
+
+    // Frycook's Delight
+    public static final String FRYCOOKS_DELIGHT = "frycooks_delight";
+    public static final TagKey<Item> HAS_FISH_SLICE = externalItemTag(FRYCOOKS_DELIGHT, "has_fish_slice");
+    public static final TagKey<Item> FISH_SLICES = externalItemTag(FarmersDelight.MODID, "fish_slices");
 
     private static TagKey<Item> externalItemTag(String modId, String path) {
         return ItemTags.create(ResourceLocation.fromNamespaceAndPath(modId, path));